refactor: do not write seed direction in init malware
When infecting a machine, this exploit is executed.
We deleted the line which writes name and directions
of the seed from which we downloaded the files.
Otherwise, a victim machine could end with a peer file
with just seed directions (since they can come from
the seed list that is downloaded).