feature: malware files can be updated
Now, we can update a file. After that we can upload
it into a seed machine for example.
In this moment we will we able to broadcast a
message to our all neighbours, that will themselves
broadcast the message to their neighbours.
All victims will end up updating the malware.