refactor: remove malware.py
This file is no longer being used for the demo.

At the beginning we directly executed it on the
victim machine, but since we figured out how to
combine the vim and sudo exploit. We decided to
design a special file to infect the victim.

The file is /scripts/exploit